+++ crates/release_plz/tests/all/changelog.rs
@@ -4,7 +4,7 @@ use git_cmd::Repo;
 use release_plz_core::{GitBackend, GitClient, Gitea, RepoUrl};
 use secrecy::SecretString;
 
-use crate::helpers::gitea_client::GiteaContext;
+use crate::helpers::gitea::GiteaContext;
 
 #[tokio::test]
 async fn release_plz_adds_changelog_on_new_project() {
