+++ crates/release_plz/tests/all/helpers/mod.rs
@@ -1,3 +1,3 @@
-mod gitea_client;
-mod gitea;
+mod context;
+pub mod gitea;
 mod reqwest_utils;
