+++ crates/release_plz/tests/all/main.rs
@@ -1,2 +1,3 @@
+mod changelog;
 mod completion_test;
 mod helpers;
